check_invitation_token_and_team_capacityをモジュールに切り出す
